Reformatting dependencies
- Deprecated Rarity.py, moved to Helpers to avoid circular import issues
- Moved get_hierarchy() function from DNA_Generator.py to Helpers to avoid circular import issues
- Reformatted some comments and example code
